requirements.txt
pymysql==1.0.3
pydantic==1.10.7
sqlalchemy==2.0.10
requests==2.28.2
beautifulsoup4==4.12.2
PyYAML==6.0